+++ api/src/resources/dataset.rs
@@ -18,6 +18,8 @@ use std::{
     str::FromStr,
 };
 
+use super::validation::ModelRating;
+
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
 #[serde(rename_all(serialize = "snake_case", deserialize = "snake_case"))]
 pub enum DatasetFlag {
@@ -49,13 +51,15 @@ pub struct Dataset {
     pub dataset_flags: Vec<DatasetFlag>,
 }
 
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
 pub struct DatasetStats {
-    pub num_reviewed: NotNan<f64>,
+    pub num_reviewed: usize,
     pub total_verbatims: NotNan<f64>,
+    pub model_rating: ModelRating,
+    pub latest_model_version: ModelVersion,
 }
 
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
 pub struct DatasetAndStats {
     pub dataset: Dataset,
     pub stats: DatasetStats,
@@ -202,6 +206,12 @@ pub struct ModelFamily(pub String);
 
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Hash)]
 pub struct ModelVersion(pub u32);
+
+impl std::fmt::Display for ModelVersion {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
 
 // TODO(mcobzarenco)[3963]: Make `Identifier` into a trait (ensure it still implements
 // `FromStr` so we can take T: Identifier as a clap command line argument).